Refactored into 2 maven modules - one for client and one for service
