--- src/main/asciidoc/cheatsheet/UpdateOptions.adoc
@@ -1,22 +1,0 @@
-== UpdateOptions
-
-++++
- Options for configuring updates.
-++++
-'''
-
-[cols=">25%,^25%,50%"]
-[frame="topbot"]
-|===
-^|Name | Type ^| Description
-
-|[[upsert]]`upsert`
-|`Boolean`
-|+++
-Set whether upsert is enabled+++
-
-|[[multi]]`multi`
-|`Boolean`
-|+++
-Set whether multi is enabled+++
-|===
