--- src/test/java/io/vertx/ext/mongo/MongoClientTestBase.java
@@ -1,93 +1,0 @@
-/*
- * Copyright 2014 Red Hat, Inc.
- *
- *  All rights reserved. This program and the accompanying materials
- *  are made available under the terms of the Eclipse Public License v1.0
- *  and Apache License v2.0 which accompanies this distribution.
- *
- *  The Eclipse Public License is available at
- *  http://www.eclipse.org/legal/epl-v10.html
- *
- *  The Apache License v2.0 is available at
- *  http://www.opensource.org/licenses/apache2.0.php
- *
- *  You may elect to redistribute this code under either of these licenses.
- */
-
-package io.vertx.ext.mongo;
-
-import de.flapdoodle.embed.mongo.MongodExecutable;
-import de.flapdoodle.embed.mongo.MongodStarter;
-import de.flapdoodle.embed.mongo.config.IMongodConfig;
-import de.flapdoodle.embed.mongo.config.MongodConfigBuilder;
-import de.flapdoodle.embed.mongo.config.Net;
-import de.flapdoodle.embed.mongo.distribution.Version;
-import de.flapdoodle.embed.process.runtime.Network;
-import io.vertx.core.json.JsonObject;
-import io.vertx.test.core.VertxTestBase;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-
-/**
- * @author <a href="http://tfox.org">Tim Fox</a>
- */
-public abstract class MongoClientTestBase extends VertxTestBase {
-
-  protected static String getConnectionString() {
-    return getProperty("connection_string");
-  }
-
-  protected static String getDatabaseName() {
-    return getProperty("db_name");
-  }
-
-  protected static String getProperty(String name) {
-    String s = System.getProperty(name);
-    if (s != null) {
-      s = s.trim();
-      if (s.length() > 0) {
-        return s;
-      }
-    }
-
-    return null;
-  }
-
-  protected JsonObject getConfig() {
-    JsonObject config = new JsonObject();
-    String connectionString = getConnectionString();
-    if (connectionString != null) {
-      config.put("connection_string", connectionString);
-    } else {
-      config.put("connection_string", "mongodb://localhost:27018");
-    }
-    String databaseName = getDatabaseName();
-    if (databaseName != null) {
-      config.put("db_name", databaseName);
-    }
-    return config;
-  }
-
-  private static MongodExecutable exe;
-
-  @BeforeClass
-  public static void startMongo() throws Exception {
-    if (getConnectionString() == null ) {
-      IMongodConfig config = new MongodConfigBuilder().
-        version(Version.Main.PRODUCTION).
-        net(new Net(27018, Network.localhostIsIPv6())).
-        build();
-      exe = MongodStarter.getDefaultInstance().prepare(config);
-      exe.start();
-    }
-  }
-
-  @AfterClass
-  public static void stopMongo() {
-    if (exe != null) {
-      exe.stop();
-    }
-  }
-
-
-}
