+++ vertx-mongo-client/src/main/java/examples/Examples.java
