+++ vertx-mongo-client/src/main/java/examples/package-info.java
