+++ vertx-mongo-client/src/main/java/io/vertx/ext/mongo/WriteOption.java
