+++ vertx-mongo-client/src/main/java/io/vertx/ext/mongo/impl/codec/json/AbstractJsonCodec.java
