+++ vertx-mongo-client/src/main/java/io/vertx/ext/mongo/impl/config/SSLSettingsParser.java
