+++ vertx-mongo-client/src/test/java/io/vertx/ext/mongo/MongoClientTest.java
@@ -1,0 +1,28 @@
+package io.vertx.ext.mongo;
+
+import io.vertx.core.json.JsonObject;
+
+import java.util.concurrent.CountDownLatch;
+
+/**
+ * @author <a href="http://tfox.org">Tim Fox</a>
+ */
+public class MongoClientTest extends MongoClientTestBase {
+
+  @Override
+  public void setUp() throws Exception {
+    super.setUp();
+    JsonObject config = getConfig();
+    mongoClient = MongoClient.createNonShared(vertx, config);
+    CountDownLatch latch = new CountDownLatch(1);
+    dropCollections(latch);
+    awaitLatch(latch);
+  }
+
+  @Override
+  public void tearDown() throws Exception {
+    mongoClient.close();
+    super.tearDown();
+  }
+
+}
