+++ vertx-mongo-service/src/main/resources/io.vertx.mongo-service.json
@@ -1,0 +1,11 @@
+{
+  "main": "io.vertx.ext.mongo.MongoServiceVerticle",
+  "options": {
+    "worker": false,
+    "config": {
+      "address": "vertx.mongo",
+      "connection_string": "mongodb://localhost:27017",
+      "db_name": "default_db"
+    }
+  }
+}
